reorganized files, updated quarter
